+++ internal/use_case/sign_out_all.go
@@ -1,8 +1,10 @@
 package use_case
 
 import (
+	"fmt"
 	"github.com/jmoiron/sqlx"
 	"github.com/rs/zerolog/log"
+	"wakarimi-authentication/internal/errors"
 )
 
 func (u UseCase) SignOutAll(accountID int) (err error) {
@@ -24,6 +26,17 @@ func (u UseCase) SignOutAll(accountID int) (err error) {
 }
 
 func (u UseCase) signOutAll(tx *sqlx.Tx, accountID int) (err error) {
+	isAccountExists, err := u.accountService.IsExists(tx, accountID)
+	if err != nil {
+		log.Error().Err(err).Int("accountId", accountID).Msg("Failed to check account existence")
+		return err
+	}
+	if !isAccountExists {
+		err = errors.NotFound{EntityName: fmt.Sprintf("account with id=%d", accountID)}
+		log.Error().Err(err).Int("accountId", accountID).Msg("Account not found")
+		return err
+	}
+
 	err = u.refreshTokenService.DeleteByAccount(tx, accountID)
 	if err != nil {
 		log.Error().Err(err).Int("accountId", accountID).Msg("Failed to delete account's refresh tokens")
